Support older libvirt address retrieval
Libvirt releases before 1.2.8 do not fully support the
virNetworkGetDHCPLeases used by dhcp_leases, so provide a compatibility
check to determine the version of the library in order to determine
which method to use to retrieve the mac address and then replace the
current addresses method reference with this.

Use method introspection to confirm that ruby-libvirt was compiled
against a libvirt with support to call dhcp_leases and also that the
library version of the connection is recent enough to support the
subsequent call.

Fixes: #27