Allow update_display on VMs with non-auto keymap
If the vnc keymap has been changed from auto to, for example, en-us,
calling update_display will fail with a "Call to
virDomainUpdateDeviceFlags failed: internal error cannot change keymap
setting on vnc graphics" error. This commit changes it so that if no
keymap has been passed explicitly in the options we will use the keymap
that is currently used.