Avoid querying for all networks by default
Given the network attribute of the nic is assumed to exist, provide this
as a filter attribute to skip querying libvirt for all networks and
instead request only the one relevant.

This can help avoid a race condition where multiple sets of vagrant
machines are being run on the same host using different networks where
during delete/destroy the networks of one are removed and yet at the
same time the other environment may attempt to access the attributes.

Relates-to: vagrant-libvirt/vagrant-libvirt#1438