Drop unused json, fog-json and fog-xml dependencies & add nokogiri
This gem doesn't use JSON at all and uses nokogiri directly instead of
fog-xml. This removes the dependencies and instead adds nokogiri
directly.

fog-xml has nokogiri 1.5.11 as the lower bound. This is updated to 1.11
because that version introduced native gems for 2.7. It's also already 3
years old so it should be usable for everyone. There were plenty of CVEs
in older versions.