Update default video device to vga
For over a decade the cirrus driver has been considered outdated and the
vga driver is preferred. A common alternative is qxl, but that implies
the use of spice which isn't shipped in at least RHEL.

It leaves out the vram and relies on libvirt to set a sane default.

Link: https://www.kraxel.org/blog/2014/10/qemu-using-cirrus-considered-harmful/