Deal with missing volume_type
LVM volumes do not have a volume type and after a reload the value is
set to nil. 0277e9b42589836134763dd182fe6025bf8349df introduced that
reload after saving. The mock driver doesn't properly show this
behavior, which is why it was missed. A test is added to simulate the
real world behavior.

Fixes: 0277e9b42589836134763dd182fe6025bf8349df ("Correctly load volumes after creation and cloning")
Link: https://libvirt.org/formatstorage.html#storage-volume-target-elements
Link: https://libvirt.org/storage.html#valid-logical-volume-format-types